+++ go-peer_test.go
@@ -4,6 +4,7 @@ import (
 	"bytes"
 	"os"
 	"regexp"
+	"strings"
 	"testing"
 )
 
@@ -16,17 +17,24 @@ func TestGoPeerVersion(t *testing.T) {
 		return
 	}
 
-	re := regexp.MustCompile(`##\s+(v\d+\.\d+\.\d+)\s+`)
+	re := regexp.MustCompile(`##\s+(v\d+\.\d+\.\d+~?)\s+`)
 	match := re.FindAllStringSubmatch(string(changelog), -1)
 	if len(match) < 2 {
 		t.Error("versions not found")
 		return
 	}
 
-	// current version is always previous version in the changelog
-	if match[1][1] != CVersion {
-		t.Error("the versions do not match")
-		return
+	if strings.HasSuffix(CVersion, "~") {
+		if match[0][1] != CVersion {
+			t.Error("the versions do not match")
+			return
+		}
+	} else {
+		// current version is always previous version in the changelog
+		if match[1][1] != CVersion {
+			t.Error("the versions do not match")
+			return
+		}
 	}
 
 	if match[0][1] == match[1][1] {