+++ pkg/network/conn/doc.go
@@ -1,20 +1,11 @@
 // Package conn makes it possible to hide the structure of the true message by encrypting and adding random bytes.
 //
 /*
-	NETWORK MESSAGE FORMAT
+	NETWORK CONN FORMAT
 
-	E( LEM || LV || H(LEM||LV) || H(EM||V) ) || EM || V
+	L(M) || M
 	where
-		LEM = L(EM)
-		LV  = L(V)
-		EM  = E(M)
-		where
-			E - encrypt (cipher-key)
-			H - hmac (auth-key)
-			L - length
-			M - message bytes
-			V - void bytes
-
-	Scheme: https://github.com/number571/go-peer/blob/master/images/go-peer_layer1_net_message.jpg
+		L - length (uint64)
+		M - message bytes
 */
 package conn